+++ packages/jest-config/src/getMaxWorkers.ts
@@ -5,8 +5,18 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {cpus} from 'os';
+import {
+  // @ts-expect-error - added in Node 19.4.0
+  availableParallelism,
+  cpus,
+} from 'os';
 import type {Config} from '@jest/types';
+
+function getNumCpus(): number {
+  return typeof availableParallelism === 'function'
+    ? availableParallelism()
+    : cpus()?.length ?? 1;
+}
 
 export default function getMaxWorkers(
   argv: Partial<
@@ -22,8 +32,7 @@ export default function getMaxWorkers(
     return parseWorkers(defaultOptions.maxWorkers);
   } else {
     // In watch mode, Jest should be unobtrusive and not use all available CPUs.
-    const cpusInfo = cpus();
-    const numCpus = cpusInfo?.length ?? 1;
+    const numCpus = getNumCpus();
     const isWatchModeEnabled = argv.watch || argv.watchAll;
     return Math.max(
       isWatchModeEnabled ? Math.floor(numCpus / 2) : numCpus - 1,
@@ -41,7 +50,7 @@ const parseWorkers = (maxWorkers: string | number): number => {
     parsed > 0 &&
     parsed <= 100
   ) {
-    const numCpus = cpus().length;
+    const numCpus = getNumCpus();
     const workers = Math.floor((parsed / 100) * numCpus);
     return Math.max(workers, 1);
   }